Working through demonstrating learning, bugs.
Completely removed the dependency on axl.RiskyQLearner and just copied and modified it as before. There was an issue where the base Player class was not being treated properly through two derivations, but works if derived from explicitly (super() cannot be called on super())

Both Q Learner and DQN had an issue in reset() which I think now finally, truly (FOR REAL!) ensures learning from game to game through a tournament. THIS IS NOT TEST IN A TOURNAMENT YET, but I have tested it in a series of games that is almost a tournament.

I moved static functions to a single file which will make it easier down the stretch.