10 opps, 10000 games/opp, 1000 turns/game
I have the CSVs from which this summary table was made if we need to go back to it. It takes my laptop almost an hour to re-run the base games so it's not too hard to redo if we need to.

Does not include any DQN or communication, all just base agents. Evidence of learning, considering what is optimal strategy against all of these opponents I think learning is evident.